Added support for KS/KL 3404 modules
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/trunk/pcds/epics/modules/Bx9000_MBT/current@11571 ececa60f-9f63-0410-97c7-ea05888eb260

+++ Bx9000_MBTApp/src/devBx9000Info.c
@@ -277,7 +277,7 @@ static long read_li_Bx9000Info(struct longinRecord * pli)
 		pli->val = ellCount( (ELLLIST *)&(pinforeq->pcoupler->sigptr_list) );
 		break;
 	case Bx9000Info_LI_OPTHRDID:
-		pli->val = (int)(pinforeq->pcoupler->opthread_id);
+		pli->val = (UINT32)((long)(pinforeq->pcoupler->opthread_id));
 		break;
 	}
 	pli->udf=FALSE;
